Module Zarr_lwt.Storage

module MemoryStore : sig ... end

An Lwt-aware in-memory storage backend for Zarr v3 hierarchy.

module ZipStore : Zarr.Zip.S with type 'a Deferred.t = 'a Lwt.t

An Lwt-aware Zip file storage backend for a Zarr v3 hierarchy.

module FilesystemStore : sig ... end

An Lwt-aware local filesystem storage backend for a Zarr V3 hierarchy.

module AwsStore : sig ... end

An Lwt-aware Amazon S3 bucket storage backend for a Zarr V3 hierarchy.