Module Zarr_lwt.Storage

module Deferred : Zarr.Types.Deferred with type 'a t = 'a Lwt.t

module MemoryStore : Zarr.Memory.S with module Deferred = Deferred

An Lwt-aware in-memory storage backend for Zarr v3 hierarchy.

module ZipStore : Zarr.Zip.S with module Deferred = Deferred

An Lwt-aware Zip file storage backend for a Zarr v3 hierarchy.

module HttpStore : Zarr.Http.S with module Deferred = Deferred

An Lwt-aware Http storage backend for a Zarr v3 hierarchy.

module FilesystemStore : sig ... end

An Lwt-aware local filesystem storage backend for a Zarr V3 hierarchy.

module AmazonS3Store : sig ... end

An Lwt-aware Amazon S3 bucket storage backend for a Zarr V3 hierarchy.